+++ src/app/app.module.ts
@@ -17,7 +17,8 @@ import {NpmTabComponent} from "./gui/show-code-dialog/tabs/npm-tab.component";
 import {ReactTabComponent} from "./gui/show-code-dialog/tabs/react-tab.component";
 import {VueTabComponent} from "./gui/show-code-dialog/tabs/vue-tab.component";
 import {CdnTabComponent} from "./gui/show-code-dialog/tabs/cdn-tab.component";
-import {AddAnimationDialog} from "./gui/add-animation-dialog/add-animation-dialog.component";
+import {ToastService} from "./shared/services/toast/toast.service";
+import {CodeDialogService} from "./gui/show-code-dialog/code-dialog.service";
 
 @NgModule({
   declarations: [
@@ -25,7 +26,6 @@ import {AddAnimationDialog} from "./gui/add-animation-dialog/add-animation-dialo
     TitlebarComponent,
     ParamsComponent,
     PlaygroundComponent,
-    AddAnimationDialog,
     ShowCodeDialog,
     AngularTabComponent,
     ReactTabComponent,
@@ -44,7 +44,10 @@ import {AddAnimationDialog} from "./gui/add-animation-dialog/add-animation-dialo
     NgxColorsModule,
   ],
   schemas: [CUSTOM_ELEMENTS_SCHEMA],
-  providers: [],
+  providers: [
+    CodeDialogService,
+    ToastService
+  ],
   bootstrap: [AppComponent]
 })
 export class AppModule {