+++ changelog.md
@@ -1,3 +1,8 @@
+### Version 0.1.2
+- Adding Custom Animation with file input + error handling
+- Setting the size of the component to match the size of selected background
+- Update style of inputs with nested buttons and inputs
+- Limiting background size 
 ### Version 0.1.1
 - Custom Animation can be added
 - Adjustable width of Animated Banner's parent